requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
google-cloud-storage==1.37.1
google-auth==1.28.0
